+++ changelog.md
@@ -5,6 +5,7 @@
 - a bunch of trigger utils
 - make (some) editor functions safer (remind me to do this for more of them)
 - templated easing functions and made them copy cocos easing functions
+- array utils rework
 - renamed category decl macro
 - some more editor color utils
 - better edit highway robbery